+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.0] - Unreleased
+
+### Added
+
+- Added Traceback rendering and handler
+- Added rich.constrain
+
+### Fixed
+
+- Fixed unnecessary padding
+
 ## [0.3.3] - 2020-02-04
 
 ### Fixed